VM (new pixeltower): 解释器基础

+++ include/openminecraft/vm/pixeltower/v0/om_pixeltower_frame.hpp
@@ -9,7 +9,7 @@ struct OMFrame
     OMMethod *method;
     void *stackPointer;
     void *returnAddr;
-    jint locals[0];
+    jbyte locals[0];
 };
 } // namespace openminecraft::vm::pixeltower::v0
